refactor(core): relocate rules engine into core/domain (hexagonal P4 sweep)
Pure `git mv` of regis/rules/ → regis/core/domain/rules/ + import-path
updates in the 7 importers. No behaviour change; the layered import-linter
contract stays green (the JSON Logic evaluator's playbook/utils imports are
still unlayered). First subtree of the P4 physical relocation.
