+++ regis/core/domain/rules/evaluator.py
