+++ regis/playbook/evaluator.py
@@ -12,10 +12,10 @@ from __future__ import annotations
 import logging
 from typing import Any
 
+from regis.core.domain.rules.evaluator import evaluate_rules
 from regis.playbook.context import NamedList
 from regis.playbook.presentation import resolve_presentation
 from regis.playbook.templates import _resolve_template
-from regis.rules.evaluator import evaluate_rules
 
 logger = logging.getLogger(__name__)
