+++ tests/test_oci_analyzer.py
@@ -9,7 +9,7 @@ from typing import Any
 import jsonschema
 
 from regis.analyzers.oci import OciAnalyzer, _platforms_supported
-from regis.rules.evaluator import evaluate_rules
+from regis.core.domain.rules.evaluator import evaluate_rules
 from tests.fakes import FakeImageInspector, make_ctx
 
 # ---------------------------------------------------------------------------
